+++ Sources/MapboxNavigationCore/Navigator/Internals/CoreNavigator/CoreNavigator.swift
@@ -296,15 +296,24 @@ final class NativeNavigator: CoreNavigator, @unchecked Sendable {
             return
         }
 
-        let refreshObserver = NavigatorRouteRefreshObserver(refreshCallback: { [weak self] in
-            guard let self else { return nil }
+        let refreshObserver = NavigatorRouteRefreshObserver(
+            refreshCallback: { [weak self] routeIdPrefix, routeIndex, _ in
+                guard let self else { return nil }
 
-            guard let primaryRoute = navigator.native.getPrimaryRoute() else { return nil }
-            return RouteRefreshResult(
-                updatedRoute: primaryRoute,
-                alternativeRoutes: navigator.native.getAlternativeRoutes()
-            )
-        })
+                guard let primaryRoute = navigator.native.getPrimaryRoute() else { return nil }
+                let alternativeRoutes = navigator.native.getAlternativeRoutes()
+
+                if primaryRoute.matching(routeIdPrefix: routeIdPrefix, routeIndex: routeIndex) {
+                    return .mainRoute(primaryRoute)
+                }
+                if let refreshedAlternative = alternativeRoutes.first(where: {
+                    $0.route.matching(routeIdPrefix: routeIdPrefix, routeIndex: routeIndex)
+                }) {
+                    return .alternativeRoute(alternative: refreshedAlternative)
+                }
+                return nil
+            }
+        )
         navigator.native.addRouteRefreshObserver(for: refreshObserver)
         navigator.native.startRoutesRefresh(
             forDefaultRefreshPeriodMs: UInt64(refreshPeriod * 1000),
@@ -511,4 +520,11 @@ final class NativeNavigator: CoreNavigator, @unchecked Sendable {
 enum NativeNavigatorError: Swift.Error {
     case failedToUpdateRoutes(reason: String)
     case failedToUpdateAlternativeRoutes(reason: String)
+}
+
+extension RouteInterface {
+    fileprivate func matching(routeIdPrefix: String, routeIndex: UInt32) -> Bool {
+        // TODO: (NN-3674) use full routeId instead of just prefix.
+        getRouteId().starts(with: routeIdPrefix) && getRouteIndex() == routeIndex
+    }
 }