+++ Tests/MapboxNavigationCoreTests/PredictiveCache/PredictiveCacheManagerTests.swift
@@ -26,7 +26,7 @@ final class PredictiveCacheManagerTests: XCTestCase {
         predictiveCacheConfig = PredictiveCacheConfig()
         let path = NSTemporaryDirectory()
         tileStore = TileStore.__create(forPath: path)
-        tilesetDescriptor = TilesetDescriptorFactory.build(forDataset: "", version: "")
+        tilesetDescriptor = TilesetDescriptorFactory.build(forDataset: "", version: "", includeAdas: false)
 
         predictiveCacheConfig.predictiveCacheMapsConfig.maximumConcurrentRequests = 3
         predictiveCacheConfig.predictiveCacheMapsConfig.locationConfig.currentLocationRadius = 30
@@ -45,7 +45,8 @@ final class PredictiveCacheManagerTests: XCTestCase {
             ),
             searchTilesetDescriptor: TilesetDescriptorFactory.build(
                 forDataset: "mbx-gen2",
-                version: ""
+                version: "",
+                includeAdas: false
             )
         )
         predictiveCacheConfig.predictiveCacheSearchConfig = searchConfig