+++ CHANGELOG.md
@@ -11,6 +11,18 @@
 * Fixed an issue where the `NavigationViewControllerDelegate.navigationViewController(_:didArriveAt:)` was not called for intermediate waypoints.
 * Fixed an issue where the `distanceUnit` passed to the `NavigationRouteOptions` and `NavigationMatchOptions` initializers was not correctly applied to the `distanceMeasurementSystem`.
 
+## 3.10.0-rc.1
+
+### Packaging
+
+* MapboxNavigationCore now requires [MapboxMaps v11.13.0-rc.1](https://github.com/mapbox/mapbox-maps-ios/releases/tag/v11.13.0-rc.1)
+* MapboxNavigationCore now requires [MapboxNavigationNative v324.13.0-rc.1](https://github.com/mapbox/mapbox-navigation-native-ios/releases/tag/v324.13.0-rc.1)
+
+### Routing
+
+* Fixed re-enabling route refresh after leg change.
+* Fixed route refresh of alternative routes.
+
 ## 3.10.0-beta.1
 
 ### Packaging